Implement smooth evasion

+++ src/bounding_box.rs
@@ -2,6 +2,6 @@ use crate::vec2D::*;
 
 #[derive(Copy, Clone)]
 pub struct BoundingBox{
-    pub pos: Vec2D::<i32>,
-    pub size: Vec2D::<i32>,
+    pub pos: Vec2D::<f64>,
+    pub size: Vec2D::<f64>,
 }
